fix: some erros were fixed
