+++ src/app/store/dashboard/page.tsx
@@ -1,0 +1,5 @@
+import Dashboard from "@/features/store/dashboard/components/dashboard"
+
+export default function DashboardPage() {
+  return <Dashboard />
+}
